requirements.txt
requests
python3-selenium
pyyaml